Finish (todo: add user score), begin work with notes.

+++ src/main/java/by/cascade/chatcot/connector/DeleteServlet.java
@@ -4,6 +4,8 @@ import by.cascade.chatcot.jsonmodel.PhraseDeleteJson;
 import by.cascade.chatcot.storage.databaseprocessing.DataBaseException;
 import by.cascade.chatcot.storage.databaseprocessing.phrases.PhraseAdapter;
 import by.cascade.chatcot.storage.databaseprocessing.phrases.mysql.PhrasesMySqlAdapter;
+import by.cascade.chatcot.storage.databaseprocessing.user.UserAdapter;
+import by.cascade.chatcot.storage.databaseprocessing.user.mysql.UserMySqlAdapter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -45,15 +47,21 @@ public class DeleteServlet extends HttpServlet {
      */
     private void doOperation(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, DataBaseException {
         PhraseAdapter phraseAdapter = null;
+        String userName = (String) request.getSession().getAttribute("userName");
+        UserAdapter userAdapter = new UserMySqlAdapter();
+        String userRole="";
+        if (userName!=null)
+        userRole = userAdapter.getUser(userName).getRole();
         try {
-            phraseAdapter = new PhrasesMySqlAdapter();
+            if (userName != null && userRole.equals("admin")) {
+                phraseAdapter = new PhrasesMySqlAdapter();
 
-            ObjectMapper mapper = new ObjectMapper();
-            PhraseDeleteJson phraseJson = mapper.readValue(request.getInputStream(), PhraseDeleteJson.class);
+                ObjectMapper mapper = new ObjectMapper();
+                PhraseDeleteJson phraseJson = mapper.readValue(request.getInputStream(), PhraseDeleteJson.class);
 
-            String id = phraseJson.getId();
-            phraseAdapter.deleteId(Integer.valueOf(id));
-
+                String id = phraseJson.getId();
+                phraseAdapter.deleteId(Integer.valueOf(id));
+            } else response.setStatus(401);
         } catch (DataBaseException e) {
             LOGGER.catching(e);
             throw new RuntimeException(e);
